Generalize endpoint in ClientConfig
gRPC library supports DNS as a default name-system, but others are
supported as well:
https://github.com/grpc/grpc/blob/master/doc/naming.md

This PR generalizes `ClientConfig` such that any arbitrary endpoints
are supported instead of only DNS based TCP/IP connections to a
server. The server code is unmodified as that involves lot more work;
it can be considered in a future PR.

This is a breaking change, but clients only require minimal change to
move to the new structure.